Eliminate left-recursive rules from the grammar
As we said in previous commits, there was a left-recursive production rule in our grammar and, to prevent triggering an infinite recursion, we had a temporary solution that was utilizing the `MAX_RECURSION_DEPTH` constant.

This issue is now fixed by refactoring the grammar in order to eliminate any left-recursive rules. For this case, we are rewriting the rule

```
DECLS → DECLS DECL
    | ε
```

to

```
DECLS → DECL DECLS
    | ε
```

which will solve the problem and make the code more concise, since we are consistent with the other production rules by using `self.run_rules_from_rhs` statements.