Add test for no semicolon at EOL, document found bug
In this comimt, I identified and documented a bug, along with adding a TODO. The test which is added in this commit will fail, since it tries to test the code containing the bug. I am committing this failed test on purpose, to record my insights about the code as they come up, for documentation reasons.

This test expects that the parsing will be unsuccesful, since there is no semicolon at the end of the statement. Despite that, the parsing completes successfuly.

The reason is documented in comments above the problematic code itself.