Prevent failed rules from consuming tokens
As we said in commit `fb1a779`, there was an issue with rules that would consume tokens despite failing.

In this commit, this issue is addressed by implementing the following:
* A new struct called `RuleResult` that models the result of an attempt to match a rule.
* All production rules now return an object of type `RuleResult`.
* Using the information in `RuleResult` about the number of tokens that a rule tried to consume and whether it failed or not, we can 'return' the tokens back in case of a failed attempt to match a rule. This way, we prevent any failed rule from consuming tokens.

By running `cargo test`, we can see that the failing test in commit `fb1a779` now passes.

In the future, I can refactor the code to implement a notion of consumed and peeked tokens. That is, tokens will be consumed strictly when the rule is matched, but we can peek as many tokens as we want in the future, in order to check whether a rule is mathced or not.