Define equivalent grammar without left recursive rules
In a previous commit, the grammar of the language was defined in a document, as given by the instructor of the class. That language contained some left-recursive rules, thus it could not be used to implement the parser in the code.

In this commit, an equivalent grammar is introduced, but with the left-recursive rules eliminated. This grammar will be used to model the parser in the code.