Update parser to parse expressions according to the precedence of operators
Previously, our parser was able to properly parse expressions, but it was not following any precedence rules. That is, the expression `1+2*3` would be parsed as `(1+2)*3`.

This behaviour is not necessarily bad. If the only purpose of our parser is to identify whether the syntax of a program is correct, then this is fine.

But, it would be convenient if our grammar is able to parse the expressions of a program with respect to the precedence of the operators. This can be very helpful in constructing the Abstract Syntax Tree (AST) of the given program, which in turn can help in semantic analysis and code generation.

In this commit, the rules of the parser are updated in order to follow the precedence of the operators. Some unit tests are also added. Finally, the updated Context Free Grammar is also added in "docs/grammar-with-op-precedence-without-left-recursive-rules.txt".