Build AST nodes for expressions
In this commit, we add code that builds the AST tree for expressions. Currently, it is not saved anywhere, it's just printed in the screen.

The important part of this commit is the method `construct_expression_node_from_token_range`

This method is used to construct the AST nodes for expressions. It takes a token range (start and end indices) that contains an expression and it creates an AST node that models this expression. This node is then inserted in the `token_index_to_expression_node` map.

It works in the following way:

There are 2 stacks, one storing operators and one storing operands. An operand is not necessarily a number. It can also be another expression (in the form of a Node). The code starts by iterating over the given token range.
 * If the current token is a parenthesis (either opening or closing), that token is skipped.
 * If the token is an operand (i.e. num or variable), we check if there is an entry in the `token_index_to_expression_node` map.
 * If we find an entry, it means that a previous invocation of this method parsed a subexpression starting from that token, so we skip the tokens starting from the current one all the way to the `token_end_index` stored in the map. We push the node stored in the map to the operand stack.
 * If we don't find an entry, we simply push the operand to the operand stack.
 * If the token is an operator, we simply push it to the operator stack. After filling the stacks, we construct the tree in the following way:
  1. We pop two operands and one operator from the appropriate stacks and we create a Node with these three (the parent is the operator, the children are the operands).
  2. While the operator stack is not empty, we pop one operand and one operator and, using the node previously created, we create a new node (the parent is the operator and the children are the operand and the previous node).
After building the AST node, we add it to the `token_index_to_expression_node` map.

NOTE: This method itself is 'dumb', meaning it does not understand operator precedence. All the magic happens when this method is invoked in the grammar's production rules for expressions (fn precedence_x_rule). In a bottom-up apporach, the nodes of the expressions are added to the `token_index_to_expression_node` map based on the precedence of the operators, and then the rule for the lower precedence level will understand that some subexpressions were parsed, and it will ignore them.