Implement the rest of the AST building
In a previous commit, we implemented the building of AST nodes, but only for expressions. In this commit, we extend the same logic and now the whole AST is built.

Initially, I thought that building the AST nodes for everything other than expressions would be trivial. Turns out that it is as complex as the expressions case. This happens because we have statements that support blocks of code, and this can grow recursively (for example, consider nested loops or nested if). Thus, we follow a similar approach as this commit: e23f2fe

Basically, as our parser checks the grammar rules, when a rule is matched we construct a node based on that rule. We then save this node in the `token_index_to_node` map, where the key is the index of the starting token.

Then, the next production rule that is matched may contain rules that were already matched. In this case, we check the map for these already matched rules, then we get their nodes and we construct a new node accordingly, which we also add to the map.

The code works in a bottom-up approach. That is, it starts by creating nodes for rules that are terminal, and then the higher level rules use these nodes to build new nodes and so on.

In this commit, we add some code on most of the production rule functions that simply creates the nodes based on the productions rules, and adds these nodes to the map. The root production rule `fn program_rule` adds a child node to the root AST node. This child node contains our whole AST.

This code can definitely be refactored and written more concise. I will do this in a future commnit.

Also, for now I simply clone the nodes whenever I need to add them as children. This is not optimal, as the nodes should be created only once and then simply be referenced when they are needed as children. But since my Rust skills are not good at the moment and I need to move on with the project, I do it in this silly way. I will come back and fix this problem later.