Add struct to model MIXAL mnemonics and instructions
Based on the specifications of MIXAL and MIX found here:
https://www.gnu.org/software/mdk/manual/mdk.pdf
we implement some basic structs to model the MIXAL mnemonics and instructions.

For now, we simply print some sample MIXAL instructions in the format that is expected by the linux utility `mixasm`. More information on this utility can be found here:
https://www.gnu.org/software/mdk/manual/mdk.pdf

We can then (manually, for now) copy and paste these instructions in a txt file, run `mixasm <path-to-file>` and then run `mixvm` to run the MIX computer simulator and inspect the simulator's registers, memory etc.