Fix: Construct AST expression nodes with left descendance
Before this commit, we used to construct the expression nodes of the AST with right descendance. What does this mean? Consider the expression '1 - 2 - 3'. The AST node for this expression would be:

```
  -
 / \
1   -
   / \
  2   3
```

This works fine for associative operators (like addition and multiplication) but is wrong for operators that are not (like subtraction). In the tree above, if we try and compute the value of the represented expression, we will get a wrong result.

The proper AST node that should be constructed for the expression '1 - 2 - 3' is:

```
    -
   / \
  -   3
 / \
1   2

```

The changes that we introduce in this commit fix this issue. Basically, we simply need to reverse the operand and operator stacks that we use in `construct_expression_node_from_token_range` and then make some changes on how we construct the node and its children.