Add token `Ast`
In the lexer, we add a new token `Ast(String)`. This token is not used in the language, but it serves as a grouping notion for other tokens in the AST nodes. For example, in the AST tree we use an `Ast` node in order to group together statements of the same scope.