Change type of token `Num` to `u32`
Based on the specification sheet of MIXAL and MIX, which can be found here:
https://www.gnu.org/software/mdk/manual/mdk.pdf
numbers can only be 31 bits (includes the sign bit).

This is because in MIX, a byte consists of 5 bits and a word consists of 5 bytes and a sign bit, so 31 bits in total.

The closest type that is supported by Rust is `u32`, so we update the type of `Num` token to be as consistent as possible with MIX.