Implement an initial version of the assembler
In this commit, we define and implement a starting version of the assembler, named `MixalAssembler`. To instantiate it, the following are needed:
 * An object of type `Node<usize, Token>`, which represents the AST
 * A `String` which refers to the path for the output file that will be produced. This file will contain the MIXAL assembly instructions.

In order to produce the assembly instructions using the assembler, the `run` method must be invoked.

For now, upon invoking `run`, the assembler will:
 * Use the AST to produce 2 instructions. These instructions are essential in order to start the execution of the program. The first one is the `ORIG` instruction, which "sets the value of the memory address to which following instructions will be allocated after compilation." The second one is the `END` instruction, which "Marks the end of the program. Its operand gives the start address for program execution".
 * After producing the output file, the assembler will also run the `mixasm` utility, which will translate the produced MIXAL instructions to binary instructions. The utility needs to be installed on the machine.

To test the behaviour of the produced binary instructions, you can use the `gmixvm` utility.

More information on the above instructions and the `mixasm` and `gmixvm` utilities can be found here: https://www.gnu.org/software/mdk/manual/mdk.pdf