Assembler: handle variable declaration
In this commit, we add support for handling variable declaration statements, for example `var alpha, beta, gamma, i, j, k: int;`.

Using the AST, the assembler now produces the MIXAL instructions to allocate the memory and initialize the variables with 0.

In order to do this, we need
 * The notion of a 'vtable', which will map an identifier name to its allocated memory address.
 * A pointer to the next available memory address for allocation. We can use this simple allocation strategy, since there is only one variable scope in our toy language.

We implement both of the above as variable members of the `MixalAssembler` struct.