Assembler: define enum `MixalRegister`
In this commit, we define the enum `MixalRegister`. The variants of this enum model the registers of the MIX computer. More information on the MIX registers can be found here:
https://www.gnu.org/software/mdk/manual/mdk.pdf
in section 2.1.

Note that we add the `Debug` attribute to the enum. This enables us to print the enum variants as text for debugging purposes.