Assembler: handle assignments
In this commit, we add support for handling assignment statements.

The functionality for computing expressions is not yet implemented, so we assume that the expression of the assignment is computed and stored in register RA.

After that, we look up the memory address of the assignment's identifier in the vtable, and we store the contents of register RA to that location.