Assembler: handle expressions (2/N)
In this commit, we continue the work on the assembler code that handles expressions.

We now are able to handle cases like:
 * `a = 1 + 1`
 * `a = alpha + 1`
 * `a = 1 + beta`
 * `a = alpha + beta`
for both addition and subtraction.

These are basically the simplest expressions (except expressions like `a = 1`, which are already handled), because both the operands are actual values.

An operand can be either a number or a variable, so we have 4 possible cases:
 1. Number <operator> Number
 2. Number <operator> Variable
 3. Variable <operator> Number
 4. Variable <operator> Variable
and our code should handle all these cases for every operator.

It is important to mention that arithmetic operators in MIX require the first operand to be stored in RA and the second operand in a memory address. This context is needed to understand the code in this commit.

We handle addition and subtraction together, since they can share the same functionality for loading operands to RA and to memory.

Note: addition is associative but subtraction is not. Because of that, we can optimise the generated code a little bit by splitting addition and subtraction and handling them seperately. But, for now, let's keep things simple.

Some `instruction_` methods are also implemented as needed, as well as a utility method named `token_to_arithmetic_operator_instruction_fn` which maps an operator token to an arithmetic instruction.