Assembler: handle expressions (3/N)
In this commit, we continue the work on the assembler code that handles expressions.

We now are able to handle cases like:
 * `a = 1 * 1`
 * `a = alpha * 1`
 * `a = 1 * beta`
 * `a = alpha * beta`

Multiplication shares the same functionality with addition and subtraction, but needs some extra work after the computation to properly load the result in RA.

In detail, the multiplication instruction stores the upper bits of the result in RA and the lower bits of the result in RX. For now, we don't handle overflows and only care about the lower bits of the result. Hence, we load RX to RA.

At some point, we can add code that throws an exception when the result overflows.

Reminder: our AST handles the unary minus operator as a multiplication. This means that expressions like `a = -5` are internally handled as `a = (-1) * 5`.

So, this commit also enables us to initialize variables with negative numbers (e.g. `a = -5`).

Some `instruction_` methods are also implemented as needed.