Assembler: handle expressions (4/N)
In this commit, we continue the work on the assembler code that handles expressions.

We are now able to handle cases like:
 * `a = 7 / 3`
 * `a = alpha / 3`
 * `a = 7 / beta`
 * `a = alpha / beta`
for division and modulo operators.

For context, division and modulo operators are computed with the same instruction in MIX. This instruction behaves differently than the other arithmetic operators of MIX. It uses a 10-byte dividend, where its sign and most significant bits are stored in RA, while the least significant bits are stored in RX. The divisor is stored in a memory address.

Because of the above, we need to handle the division and modulo operators together with the same approach, which differs from the addition/subtraction/multiplication method.

Some `instruction_` methods are also implemented as needed.