Assembler: handle expressions (5/N)
In this commit, we continue the work on the assembler code that handles expressions.

We refactor some of the implemented functionality in order to make the code more readable and make it easier to accommodate future operators.

Some comments and documentation are also added.

The code changes should be trivial to understand, as they simply extract some existing code to new methods.