Assembler: handle expressions (6/N)
In this commit, we continue the work on the assembler code that handles expressions.

We are now able to handle cases like:
 * `alpha = 1 == 4;` \
 * `alpha = 1 != 4;` \
 * `alpha = 1 > 4;` \
 * `alpha = 1 >= 4;` \
for every comparison operator.

The existing code already handles the preperation of the operands (ie load operand1 to RA and operand2 to an address in memory), as well as the execution of the comparison.

In this commit, we simply need to put the result of the comparison to RA.

For context, the comparison instruction in MIX (mnemonic CMPA) will set the comparison indicator register (CM) to one of the following values, based on the result of the comparison:
 * E for operand1 == operand2
 * G for operand1 > operand2
 * L for operand1 < operand2

We then use instructions like JE (jump if CM is E) etc to properly load the comparison result to RA.

Some `instruction_` and `instructions_` methods are also implemented as needed.