Assembler: handle expressions (7/N)
In this commit, we continue the work on the assembler code that handles expressions.

We are now able to handle cases like:
 * `alpha = 1 && 4;` \
 * `alpha = 1 || 4;` \
for the logical operators `&&` and `||`.

The existing code already handles the preperation of the operands (ie load operand1 to RA and operand2 to an address in memory).

In MIX, there are no instructions to perform logical `&&` and `||`, hence we build them ourselves by combining other instructions.

Some `instruction_` and `instructions_` methods are also implemented as needed.