Assembler: handle expressions (8/N)
In this commit, we continue the work on the assembler code that handles expressions.

We are now able to handle cases like:
 * `alpha = !1;`
 * `alpha = !0;`
 * `alpha = !beta;`
for the logical operator NOT (`!`).

This operator is special, since it is a unary operator. Thus, we handle it in the beginning of the `handle_expression_node` method.

The code is simple. We recursively calculate the expression on which the operator `!` is applied, and we store it in RA. We then check if RA is 0. If true, we set register RA to 1. If false, we set register RA to 0.