Assembler: handle expressions (9/9)
In this commit, we finish the work on the assembler code that handles expressions.

We are now able to handle expressions with more than one operands, for example:
 * `alpha = 1 + 2 * 3 / 4 + (5 || 6 && 7);`

Initially, I thought that it was possible to use exclusively CPU registers (without leveraging extra memory) in order to evaluate any expression. But this is not possible for every expression, because at some point it is required to store intermediate values (ie evaluated left node) while evaluating the other node. But, the latter node might also need to store intermediate values and so on.

To put it simpler, registers only give us O(1) space, which is not enough to calculate an arithmetic expression with an arbitrary number of terms. To solve this, we need to utilize memory. This is what we implement in this commit.

Note: we can still calculate some simple expressions using only registers, altough the code in this commit will always use the memory approach explained above in order to keep things simple.

The code works by executing the following steps when one or both of the operands are expressions:
 1. Evaluate expression in right operand and store it in register RA
 2. Allocate memory address A
 3. Store register RA to memory address A
 4. Evaluate expression in left operand and store it in register RA
 5. Evaluate the operator result between RA (left operand) and the memory address A (right operand). In case of division or modulo, we first handle the left operand accordingly
 6. De-allocate memory address A
 7. As we already do, execute any further instructions required in order to properly load the operator result to register RA