Assembler: handle `while` loop
In this commit, we add code that handles `while` loops.

Again, since the assembler can already handle expressions, adding support for `while` loops is fairly straightforward.

1. Run `handle_expression_node` to evaluate the condition of the `while` loop.
2. If the condition evaluates to 0, jump to step 5, else continue executing step 3.
3. Execute the block of code associated with the `while` loop.
4. Jump to step 1.
5. NOP - placeholder instruction.

In the code, we use instruction labels to make the jumps. We need 2 labels per loop:
* One label to jump to the start of the loop, prior to checking the condition, and
* One label to jump to the end of the loop, after the last instruction

In order to support `continue` and `break` statements, we maintain a stack. When we start handling the loop, we push to this stack the instruction labels that were explained above. At any given time, the top of the stack contains the instruction labels of the last-encountered loop. If we read a `continue` or `break` statement in our AST, we read the top of the stack and
* For `continue` statements: jump to the start-of-the-loop label,
* For `break` statements: jump to the end-of-the-loop label

After we are done handling the loop, we pop the last value from the stack.