Assembler: allow `handle_root` method to be invoked with any node as argument
Up to now, the root method of the assembler, namely `handle_root`, implicitly required the argument node to have a `Token::Ast` value.

We can see this in the code: the method skips the given node and directly starts handling the children nodes.

This behaviour does not generalize well, since we may need to invoke `handle_root` on any node, not just nodes with `Token::Ast` value.

Thus, we refactor the code a little bit to allow this method to be invoked with any type of node. In the case of a `Token::Ast` node, we then start invoking the same method on the children.