Assembler: handle `for` loop
In this commit, we add code that handles `for` loops.

Again, since the assembler can already handle expressions, adding support for `for` loops is fairly straightforward.

1. Run the first part of the `for` loop.
2. Run `handle_expression_node` to evaluate the second part (basically the condition) of the `for` loop.
3. If the condition evaluates to 0, jump to step 7, else continue executing step 4.
4. Execute the block of code associated with the `for` loop.
5. Execute the third part of the `for` loop.
6. Jump to step 2.
7. NOP - placeholder instruction.

In the code, we use instruction labels to make the jumps. We need 3 labels per loop:
1. One label to jump to the start of the loop, prior to checking the condition,
2. One label to jump to the end of the loop, after the last instruction,
3. One label to jump to the start of execution of the third part of the loop, which must be executed at the end of every iteration

Similarly to the commit where we handled the `while` loop (8dbd9b6), for every loop we need to push 2 labels to the stack, one corresponding to `continue` statements and one corresponding to `break` statements.
* For `continue` statements, we need to push to the stack the 3rd label that we explained above, in order to ensure that the third part of the loop executes, even if we `continue`.
* For `break` statements, we push to the stack the 2nd label that we explained above, similarly to the `while` loop case.