Assembler: perform short-circuit evaluation for operators `&&` and `||`
In this commit, we add code that performs short-circuit evaluation for the logical operators `&&` and `||`.

We add the methods `instructions_prepare_logical_and_operands` and `instructions_prepare_logical_or_operands`, which first evaluate the left operand of the operator, and, based on the result, decide whether the right operand should be evaluated or not. Then, the appropriate values are entered to register RA and the `address` memory location to prepare for the evaluation of the operator.

Notice that the logic for the evaluation of the operators is not changed in any way, just the logic that prepares the operands of the operator.