Assembler: throw exception when dividing with 0
In this commit, we add code that prints an exception to the terminal and halts the program when a zero division is encountered.

We first rename the method `instruction_divide_and_modulo` to `instructions_divide_and_modulo`, because we will be adding more assembly intrustions in its body.

We then add code that handles the case of zero division, prior of the division execution.

When a zero division is encountered, the method `instructions_throw_exception(exception_code)` is invoked. We pick number `1` as the code for this exception.

Since we haven't implemented any sophisticated exception handling yet, for now we simply print the exception code to the terminal and halt the execution of the program.