Tooling: Run the compiler for a source code file provided with command line args
In this commit, we work on the tooling of the compiler.

First of all, our dummy toy language now has a name: YAL (stands for Yet Another Language). Starting from now, the source code files that we test with will have a `.yal` file extension.

In this commit:
* We create the struct `FilesHandler`. This struct is initialized with a YAL source code file path, then it generates and stores all the names of files and directories that are used throughout the phases of the compiler. It does not create the files, it just generates the names that will be used. Then, other parts of the compiler will read the names from the `FilesHandler` struct and will create the files accordingly. Note that it will create new directories if needed. For now, we store the output files that are created by the compiler under a `bin` directory that we create in the same directory as the YAL source code file path.
* We add a new utility function that executes an operating system command for running a MIX binary file. It uses the `mixvm` utility to run the binary file.
* We use the functionality explained above in `main` in order to read an input file from the command line args, get the file names that we need using a `FilesHandler` instance and then run the steps of the compiler. If the caller specifies the `--run` flag, then we also execute the MIX binary file and print its outputs.

Out compiler is now able to compile, run and print the output of a YAL program using just a command.

Finally, we add a new folder named `code-snippets` in which we store some YAL programs for testing and debugging.