Assembler: Allocate a block in memory for the standard output device
In this commit, we work on allocating a fixed memory block for the standard output device.

From the MIX specs, we know that the standard output device expects blocks of 14 words. That is, you specify a starting address in memory and a range of 14 words, starting from that address, is transfered to the standard output device. Then, that block of words is printed to the standard output.

Thus, we allocate the last 14 words of the address space for this purpose.

We then update the `handle_print` method to use this allocated block for executing the `print` statements.

Prior to this change, whenever we needed to execute a `print` statement, we had to allocate a new block of 14 words. The problem is: if parts of the memory were allocated and freed before the `print` statement, then they could potentially contain junk. So, we had to ensure that the 14 words block is clean (i.e. all words are set to 0) before storing the content that we wanted to print.

After this commit, we are ensured that the pre-allocated memory is 'clean' without the need to check the content of the memory on every `print`.